[middleware/checklist/blacklist/whitelist] Auditor::Result as a separated class; code review
