+++ lib/blast_wave/check_list/auditor.rb
@@ -4,30 +4,7 @@ module Rack
   # @api private
   # @since 0.1.0
   class BlastWave::CheckList::Auditor
-    # @api private
-    # @since 0.1.0
-    class Result
-      # @return [Boolean]
-      #
-      # @api private
-      # @since 0.1.0
-      attr_accessor :locked
-
-      # @return [Array<String>]
-      #
-      # @api private
-      # @since 0.1.0
-      attr_reader :triggered_filters
-
-      # @return [void]
-      #
-      # @api private
-      # @since 0.1.0
-      def initialize
-        @locked = false
-        @triggered_filters = []
-      end
-    end
+    require_relative 'auditor/result'
 
     # @param checker [BlastWave::CheckList::Checker]
     # @return [void]
